requirements.txt
flask
flask-restful
requests
Pillow
